Extract and split up appimage_handler related code
This allows for some maintenance, as code can be found more easily.

CC #1
