+++ src/libappimage/appimage_handler_type2.h
@@ -1,0 +1,6 @@
+#pragma once
+
+// local includes
+#include "appimage_handler.h"
+
+appimage_handler appimage_type_2_create_handler();
