+++ src/libappimage/libappimage.c
@@ -1930,349 +1930,6 @@ int appimage_unregister_in_system(const char *path, bool verbose)
     return 0;
 }
 
-
-
-bool is_handler_valid(const appimage_handler *handler) {
-    if (!handler) {
-#ifdef STANDALONE
-        fprintf(stderr, "WARNING: Invalid handler found, you should take a look at this now!");
-#endif
-        return false;
-    }
-
-    return true;
-}
-
-void mk_base_dir(const char *target)
-{
-    gchar *dirname = g_path_get_dirname(target);
-    if(g_mkdir_with_parents(dirname, 0755)) {
-#ifdef STANDALONE
-        fprintf(stderr, "Could not create directory: %s\n", dirname);
-#endif
-    }
-
-    g_free(dirname);
-}
-
-/*
- * Dummy fallback functions
- */
-void dummy_traverse_func(appimage_handler *handler, traverse_cb command, void *data) {
-    (void) handler;
-    (void) command;
-    (void) data;
-
-    fprintf(stderr, "Called %s\n", __FUNCTION__);
-}
-
-char* dummy_get_file_name (appimage_handler *handler, void *data) {
-    fprintf(stderr, "Called %s\n", __FUNCTION__);
-}
-
-void dummy_extract_file(struct appimage_handler *handler, void *data, char *target) {
-    fprintf(stderr, "Called %s\n", __FUNCTION__);
-}
-
-/*
- * AppImage Type 1 Functions
- */
-
-void appimage_type1_open(appimage_handler *handler) {
-    if ( is_handler_valid(handler) && !handler->is_open ) {
-#ifdef STANDALONE
-        fprintf(stderr, "Opening %s as Type 1 AppImage\n", handler->path);
-#endif
-        struct archive *a;
-        a = archive_read_new();
-        archive_read_support_format_iso9660(a);
-        if (archive_read_open_filename(a, handler->path, 10240) != ARCHIVE_OK) {
-            fprintf(stderr, "%s", archive_error_string(a));
-            handler->cache = NULL;
-            handler->is_open = false;
-        } else {
-            handler->cache = a;
-            handler->is_open = true;
-        }
-    }
-}
-
-void appimage_type1_close(appimage_handler *handler) {
-    if ( is_handler_valid(handler) && handler->is_open ) {
-#ifdef STANDALONE
-        fprintf(stderr, "Closing %s\n", handler->path);
-#endif
-        struct archive *a = handler->cache;
-        archive_read_close(a);
-        archive_read_free(a);
-
-        handler->cache = NULL;
-        handler->is_open = false;
-    }
-}
-
-void appimage_type1_traverse(appimage_handler *handler, traverse_cb command, void *command_data) {
-    appimage_type1_open(handler);
-
-    if (!command) {
-#ifdef STANDALONE
-        fprintf(stderr, "No traverse command set.\n");
-#endif
-        return;
-    }
-
-    if (handler->is_open) {
-        struct archive *a = handler->cache;
-        struct archive_entry *entry;
-        int r;
-
-        for (;;) {
-            r = archive_read_next_header(a, &entry);
-            if (r == ARCHIVE_EOF) {
-                break;
-            }
-            if (r != ARCHIVE_OK) {
-                fprintf(stderr, "%s\n", archive_error_string(a));
-                break;
-            }
-
-            /* Skip all but regular files; FIXME: Also handle symlinks correctly */
-            if(archive_entry_filetype(entry) != AE_IFREG) {
-                continue;
-            }
-
-            command(handler, entry, command_data);
-        }
-    }
-
-    appimage_type1_close(handler);
-}
-
-char* appimage_type1_get_file_name (appimage_handler *handler, void *data) {
-    (void) handler;
-
-    struct archive_entry *entry = (struct archive_entry *) data;
-
-    char *filename = replace_str(archive_entry_pathname(entry), "./", "");
-    return filename;
-}
-
-void appimage_type1_extract_file (appimage_handler *handler, void *data, const char *target) {
-    (void) data;
-
-    struct archive *a = handler->cache;
-    mk_base_dir(target);
-
-    mode_t mode = S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH;
-    int f = open(target, O_WRONLY | O_CREAT | O_TRUNC, mode);
-
-    if (f == -1){
-#ifdef STANDALONE
-        fprintf(stderr, "open error: %s\n", target);
-#endif
-        return;
-    }
-
-    archive_read_data_into_fd(a, f);
-    close(f);
-}
-
-appimage_handler appimage_type_1_create_handler() {
-    appimage_handler h;
-    h.traverse = appimage_type1_traverse;
-    h.get_file_name = appimage_type1_get_file_name;
-    h.extract_file = appimage_type1_extract_file;
-
-    return h;
-}
-
-/*
- * AppImage Type 2 Functions
- */
-
-void appimage_type2_open(appimage_handler *handler) {
-    if (is_handler_valid(handler) && !handler->is_open) {
-#ifdef STANDALONE
-        fprintf(stderr, "Opening %s as Type 2 AppImage\n", handler->path);
-#endif
-        // The offset at which a squashfs image is expected
-        ssize_t fs_offset = appimage_get_elf_size(handler->path);
-
-        if (fs_offset < 0) {
-#ifdef STANDALONE
-            fprintf(stderr, "get_elf_size error\n");
-#endif
-            handler->is_open = false;
-            handler->cache = NULL;
-            return;
-        }
-
-        sqfs *fs = malloc(sizeof(sqfs));
-        sqfs_err err = sqfs_open_image(fs, handler->path, (size_t) fs_offset);
-        if (err != SQFS_OK){
-#ifdef STANDALONE
-            fprintf(stderr, "sqfs_open_image error: %s\n", handler->path);
-#endif
-            free(fs);
-            handler->is_open = false;
-            handler->cache = NULL;
-        } else {
-            handler->is_open = true;
-            handler->cache = fs;
-        }
-    }
-}
-
-void appimage_type2_close(appimage_handler *handler) {
-    if ( is_handler_valid(handler) && handler->is_open ) {
-#ifdef STANDALONE
-        fprintf(stderr, "Closing %s\n", handler->path);
-#endif
-
-        sqfs_destroy(handler->cache);
-        free(handler->cache);
-
-        handler->is_open = false;
-        handler->cache = NULL;
-    }
-}
-
-void appimage_type2_traverse(appimage_handler *handler, traverse_cb command, void *command_data) {
-    appimage_type2_open(handler);
-
-    if (handler->is_open && handler->cache != NULL) {
-        sqfs *fs = handler->cache;
-        sqfs_traverse trv;
-        sqfs_inode_id root_inode = sqfs_inode_root(fs);
-        sqfs_err err = sqfs_traverse_open(&trv, fs, root_inode);
-        if (err!= SQFS_OK) {
-#ifdef STANDALONE
-            fprintf(stderr, "sqfs_traverse_open error\n");
-#endif
-        }
-        while (sqfs_traverse_next(&trv, &err))
-            command(handler, &trv, command_data);
-
-        if (err) {
-#ifdef STANDALONE
-            fprintf(stderr, "sqfs_traverse_next error\n");
-#endif
-        }
-        sqfs_traverse_close(&trv);
-    }
-
-    appimage_type2_close(handler);
-}
-
-char* appimage_type2_get_file_name (appimage_handler *handler, void *data) {
-    (void) handler;
-    sqfs_traverse *trv = data;
-    return strdup(trv->path);
-}
-
-void appimage_type2_extract_symlink(sqfs *fs, sqfs_inode *inode, const char *target);
-
-void appimage_type2_extract_regular_file(sqfs *fs, sqfs_inode *inode, const char* target) {
-    mk_base_dir(target);
-
-    // Read the file in chunks
-    squash_extract_inode_to_file(fs, inode, target);
-}
-void appimage_type2_extract_file_following_symlinks(sqfs *fs, sqfs_inode *inode, const char* target) {
-    if(inode->base.inode_type == SQUASHFS_REG_TYPE)
-        appimage_type2_extract_regular_file(fs, inode, target);
-    else if(inode->base.inode_type == SQUASHFS_SYMLINK_TYPE) {
-        appimage_type2_extract_symlink(fs, inode, target);
-    } else {
-#ifdef STANDALONE
-        fprintf(stderr, "WARNING: Unable to extract file of type %d", inode->base.inode_type);
-#endif
-    }
-}
-
-void appimage_type2_extract_symlink(sqfs *fs, sqfs_inode *inode, const char *target) {
-    size_t size;
-    sqfs_readlink(fs, inode, NULL, &size);
-    char buf[size];
-    int ret = sqfs_readlink(fs, inode, buf, &size);
-    if (ret != 0) {
-#ifdef STANDALONE
-        fprintf(stderr, "WARNING: Symlink error.");
-#endif
-    }
-    else {
-
-        sqfs_err err = sqfs_inode_get(fs, inode, fs->sb.root_inode);
-        if (err != SQFS_OK) {
-#ifdef STANDALONE
-            fprintf(stderr, "WARNING: Unable to get the root inode. Error: %d", err);
-#endif
-        }
-
-        bool found = false;
-        err = sqfs_lookup_path(fs, inode, buf, &found);
-        if (err != SQFS_OK) {
-#ifdef STANDALONE
-            fprintf(stderr, "WARNING: There was an error while trying to lookup a symblink. Error: %d", err);
-#endif
-        }
-
-        if (found)
-            appimage_type2_extract_file_following_symlinks(fs, inode, target);
-    }
-}
-
-void appimage_type2_extract_file (appimage_handler *handler, void *data, const char *target) {
-    sqfs *fs = handler->cache;
-    sqfs_traverse *trv = data;
-
-    sqfs_inode inode;
-    if(sqfs_inode_get(fs, &inode, trv->entry.inode)) {
-#ifdef STANDALONE
-        fprintf(stderr, "sqfs_inode_get error\n");
-#endif
-    }
-
-    appimage_type2_extract_file_following_symlinks(fs, &inode, target);
-}
-
-appimage_handler appimage_type_2_create_handler() {
-    appimage_handler h;
-    h.traverse = appimage_type2_traverse;
-    h.get_file_name = appimage_type2_get_file_name;
-    h.extract_file = appimage_type2_extract_file;
-
-    return h;
-}
-
-/* Factory function for creating the right appimage handler for
- * a given file. */
-appimage_handler create_appimage_handler(const char *const path) {
-    int appimage_type = appimage_get_type(path, 0);
-
-    appimage_handler handler;
-#ifdef STANDALONE
-    fprintf(stderr,"AppImage type: %d\n", appimage_type);
-#endif
-    switch (appimage_type) {
-        case 1:
-            handler = appimage_type_1_create_handler();
-            break;
-        case 2:
-            handler = appimage_type_2_create_handler();
-            break;
-        default:
-#ifdef STANDALONE
-            fprintf(stderr,"Appimage type %d not supported yet\n", appimage_type);
-#endif
-            handler.traverse = dummy_traverse_func;
-            break;
-    }
-    handler.path = path;
-    handler.is_open = false;
-    return handler;
-}
-
 void move_file(const char *source, const char *target) {
     g_type_init();
     GError *error = NULL;
